wip: better impl
