+++ src/checkstyle/suppressions.xml
@@ -34,10 +34,10 @@
     <suppress checks="JavadocPackage" files="com.flowlogix.examples.*"/>
     <suppress checks="JavadocType" files="PackageMarker"/>
     <suppress checks="JavadocType" files="InternalQualifierJPALazyModel" lines="26"/>
-    <suppress checks="JavadocType" files="JPAModelImpl" lines="84,166,203"/>
+    <suppress checks="JavadocType" files="JPAModelImpl" lines="85,178,215"/>
     <suppress checks="OuterTypeNumber" files="CursorPagination" lines="16"/>
     <suppress checks="JavadocType" files="CursorPagination" lines="141"/>
-    <suppress checks="JavadocType" files="JPALazyDataModel.java" lines="145"/>
+    <suppress checks="JavadocType" files="JPALazyDataModel.java" lines="135"/>
 
     <suppress checks="JavadocVariable" lines="49-52" files="com.flowlogix.util.ShrinkWrapManipulator"/>
     <suppress checks="ConstantName" lines="60-63" files="com.flowlogix.util.TypeConverter"/>
